[VACGOM-90] refactor: extractToken 메소드 책임 이동
